Separated string, statistics, and bitwise operations to separate files and created unit tests
