+++ .gitignore
@@ -1,3 +1,3 @@
-# excludes all pycache files
-methods/__pycache__/
-__pycache__/
+# prevents all pycaches from being added 
+__pycache__
+easyPythonpi/__pycache__
