+++ easyPythonpi/test/test_basics.py
@@ -3,7 +3,7 @@
 
 import unittest
 
-from easyPythonpi.methods.basics import *
+from methods.basics import *
 
 class TestBasicMath(unittest.TestCase):
     def test_add_positive_numbers(self):
